aplicacao/sentiment-analysis-api/requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.96.0
filelock==3.0.12
flake8==3.8.4
gdown==3.12.2
h11==0.12.0
idna==2.10
isort==5.7.0
joblib==1.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
pathspec==0.8.1
pycodestyle==2.6.0
pydantic==1.10.9
pyflakes==2.2.0
pyparsing==2.4.7
PySocks==1.7.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
six==1.15.0
starlette==0.27.0
tokenizers==0.10.1
toml==0.10.2
torch==1.8.0
tqdm==4.59.0
transformers==4.3.3
typed-ast==1.4.2
typing-extensions==4.6.3
urllib3==1.26.3
uvicorn==0.22.0
openai==0.27.7
python-dotenv==1.0.0